+++ CHANGELOG.md
@@ -134,6 +134,7 @@
 - Use constant for max list count instead of 500000 (#352)
 - Tighten wildcard searches to only search in the same domain level (#348)
 - useless conditional when fetching id parameter (#353)
+- ES 5.0 no longer supports the write consistency option for index(); archiver fails (#351)
 
 
 ## CHANGES in 0.9b: